x/eav/types/message_create_new_user_test.go
package types

import (
	"testing"

	"belShare/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateNewUser_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateNewUser
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateNewUser{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateNewUser{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
